mavlink: move MANUAL_CONTROL to separate streams header
